[refactor] 연결 기능 connect 도메인으로 분리
